Read over it
I changed 2b. To me it sounded, as if we did not understand what we
write, then I think it’s better to call the problems we see by name.
Change it as you want and then submit, I am ok with it. Just put me in
the CC. And thanks for the good work this semester!